Make dump()-method public without parameters
The dump()-method now uses private attributes instead of parameters and it is public now. The constructor does not call the dump()-method anymore, but only sets the private attributes. The dump()-method now has to be called from outside the DumpTask class.

Also, the DumpTask class now uses the DriverManager namespace.