Refactor output-format related code into a dedicated class + interface (#88)
We are looking into adding an option to do CSV exports as well. As a first step in that direction, we need to separate everything related to how we write export SQL from things related to fetching the data and user interface. Currently, that's all mixed up in the `Dumper` class.

This PR takes a first step in that direction by introducing the `OutputFormatDriverInterface`, and moving everything related to exporting SQL from the `Dumper` class to `MysqlOutputFormatDriver`.

Co-authored-by: Matthias Pigulla <mp@webfactory.de>