Fill in dummy types to deal with DC2:... column comments (#102)
We're using the `SchemaManager` to simplify listing tables, views and so
on. But, the information provided by it also includes the DBAL types
used for columns, and that information may be parsed from column
comments in the database.

The problem is that a table definition like so:

```sql
CREATE TABLE `test` (
  `some_col` int NOT NULL COMMENT '(DC2Type:some_type)'
)
```

... `slimdump` fail fail with an exception because it does not know
about the `some_type`. This type may be available and defined in your
application, but we cannot put it into the `slimdump` command.

Now, in fact, we don't care about the type and its conversion
capabilities at all. So, this PR adds a runtime "just in type"
registration of a `DummyType` that serves as a placeholder to avoid
issues when loading the schema.

Fixes #96.